backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
xhtml2pdf==0.2.17
gunicorn==21.2.0
whitenoise==6.6.0
django-ratelimit==4.1.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
